feat: update script for new versions
